Initial commit of the Batch sample
This commit creates a simple example of using a function to provide the
logic behind a Spring Batch ItemReader, ItemProcessor, and ItemWriter.
In order to get this to work, I started by hand coding functions (as
seen in the SampleApplication class).  The next step I took was I tried
to get them to be picked up by the registry so that the Spring Batch
components could get them from there.  However, since I am not returning
Flux, they were not picked up by the
ContextFunctionCatalogAutoConfiguration (It seems to only pick up
functions that deal with Flux).  I tweaked the
ContextFunctionCatalogAutoConfiguration to drop the Flux requirement but
that breaks tests.  Finally, I tried configuring dynamicly compiled
functions in the applicaiton.yml file but the supplier keeps returning a
FluxArray which is a package private class so I can't do anything with
it.